基本的なコメント機能の実装
