+++ app/assets/stylesheets/_posts.scss
@@ -131,6 +131,11 @@ a {
   }
 }
 
+button, input, select, textarea {
+  font-family : inherit;
+  font-size   : 100%;
+}
+
 .button {
   &-post {
     text-align: left;
@@ -152,6 +157,48 @@ a {
       color: white;
     }
   }
+  &-comment {
+    color: white;
+    border: none;
+    outline: none;
+    background-color: #b3642f;;
+    border-radius: 30px;
+    padding: 3px 8px;
+    transition: .1s;
+    &:hover {
+      background-color: #dfa57e;
+      cursor: pointer;
+    }
+  }
+}
+
+
+%post-comment-header {
+  display: flex;
+  margin-bottom: 5px;
+}
+%post-comment-header__name {
+  font-size: 1.1rem;
+  font-weight: bold;
+  margin-right: 20px;
+}
+%post-comment-header__time {
+  font-size: 0.9rem;
+  color: #6d6d6d;
+  margin-right: 10px;
+}
+%post-comment-header__delete {
+  color: #6d6d6d;
+}
+%post-comment-text {
+  margin-bottom: 10px;
+}
+%post-comment-image {
+  border-radius: 5%;
+  height: 100px;
+  width: 100px;
+  object-fit: cover;
+  margin-bottom: 10px;
 }
 
 .post {
@@ -171,20 +218,15 @@ a {
     width: 100%;
   }
   &-header {
-    display: flex;
-    margin-bottom: 5px;
+    @extend %post-comment-header;
     &__name {
-      font-size: 1.1rem;
-      font-weight: bold;
-      margin-right: 20px;
+      @extend %post-comment-header__name;
     }
     &__time {
-      font-size: 0.9rem;
-      color: #6d6d6d;
-      margin-right: 10px;
+      @extend %post-comment-header__time;
     }
     &__delete {
-      color: #6d6d6d;
+      @extend %post-comment-header__delete;
     }
   }
   &-review {
@@ -204,19 +246,16 @@ a {
     }
   }
   &-text {
-    margin-bottom: 10px;
+    @extend %post-comment-text;
   }
   &-image img {
-    border-radius: 5%;
-    height: 100px;
-    width: 100px;
-    object-fit: cover;
-    margin-bottom: 10px;
+    @extend %post-comment-image;
   }
   &-action {
     width: 100%;
     display: flex;
     justify-content: flex-end;
+    margin-bottom: 5px;
     i {
       font-size: 1.1rem;
     }
@@ -233,21 +272,84 @@ a {
       }
     }
   }
-  &-comment {
-    &-list {
-      display: flex;
-      margin: 10px 0;
-      img {
-        border-radius: 25%;
-        width: 50px;
-        height: 50px;
-        border: 1px solid #ffffff;
-      }
-    }
-    &-left {
-      margin-right: 10px;
+}
+
+.comment {
+  &-list, &-write {
+    display: flex;
+    padding: 10px 0;
+    border-top : solid 1px #a1a1a1;
+  }
+  &-left, &-write-left {
+    margin-right: 10px;
+    img {
+      border-radius: 25%;
+      width: 50px;
+      height: 50px;
+      border: 1px solid #ffffff;
     }
   }
+  &-right, &-write-right {
+    width: calc(100% - 170px);
+    box-sizing: border-box;
+  }
+  &-header {
+    @extend %post-comment-header;
+    &__name {
+      @extend %post-comment-header__name;
+    }
+    &__time {
+      @extend %post-comment-header__time;
+    }
+    &__delete {
+      @extend %post-comment-header__delete;
+    }
+  }
+  &-text {
+    @extend %post-comment-text;
+  }
+  &-image img {
+    @extend %post-comment-image;
+  }
+  &-write {
+    &__textarea {
+      box-sizing: border-box;
+      width: 100%;
+      textarea {
+        box-sizing: border-box;
+        width: 100%;
+        height: 80px;
+        padding: 5px;
+        border : solid 1px #a1a1a1;
+        resize: none;
+        &:focus {
+          outline: none;
+          background-color: #ffe0cc;
+        }
+      }
+    }
+    &__submit {
+      display: flex;
+      justify-content: flex-end;
+      &--img {
+        padding-top: 5px;
+        i {
+          font-size: 1.4rem;
+          color: #b3642f;
+          &:hover {
+            color: #dfa57e;
+            cursor: pointer;
+          }
+        }
+        margin-right:10px;
+      }
+    }
+  }
+  
+}
+
+.form-image {
+  input {display: none;}
 }
 
 hr {
