+++ app/assets/stylesheets/application.scss
@@ -14,3 +14,4 @@
 @import "font-awesome";
 @import "posts";
 @import "flash";
+@import "comments";
