+++ app/models/comment.rb
@@ -1,4 +1,9 @@
 class Comment < ApplicationRecord
+  validates :message, presence: true, unless: Proc.new {images.attached?}
+  validates :message, length: { maximum: 500 }
+  include ImageValidation
+
   belongs_to :user
   belongs_to :post
+  has_many_attached :images, dependent: :destroy
 end
