+++ app/models/post.rb
@@ -1,7 +1,7 @@
 class Post < ApplicationRecord
   validates :text, presence: true, unless: Proc.new {images.attached?}
   validates :text, length: { maximum: 500 }
-  validate :image_length, :image_type, :image_size
+  include ImageValidation
 
   belongs_to :user
   has_many_attached :images, dependent: :destroy
@@ -11,30 +11,5 @@ class Post < ApplicationRecord
   has_many :comments, dependent: :destroy
 
   attr_accessor :rev_flg
-
-  def image_length
-    if images.length > 4
-      images.purge
-      errors.add(:images, "は4枚まで同時投稿が可能です。")
-    end
-  end
-
-  def image_type
-    images.each do |image|
-      if !image.blob.content_type.in?(%('image/jpeg image/png'))
-        images.purge
-        errors.add(:images, 'はjpegまたはpng形式でアップロードしてください')
-      end
-    end
-  end
-
-  def image_size
-    images.each do |image|
-      if image.blob.byte_size > 3.megabytes
-        image.purge
-        errors.add(:images, "は1つのファイル3MB以内にしてください")
-      end
-    end
-  end
 
 end
